# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WARROOM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: getAgents
      tags: [agents]
      summary: List all agents
      parameters:
        - name: domainId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/generate:
    post:
      operationId: generateAgent
      tags: [agents]
      summary: AI-generate an agent profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expert_type:
                  type: string
                domain:
                  type: string
                domain_id:
                  type: integer
                prior_background:
                  type: string
                key_focus_area:
                  type: string
                known_bias:
                  type: string
              required: [expert_type, domain]
      responses:
        "200":
          description: Generated agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an agent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    put:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /chains:
    get:
      operationId: getChains
      tags: [chains]
      summary: List all chains
      parameters:
        - name: domainId
          in: query
          schema:
            type: integer
        - name: scenarioId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of chains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chain"
    post:
      operationId: createChain
      tags: [chains]
      summary: Create a chain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChainBody"
      responses:
        "201":
          description: Created chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chain"
  /chains/generate:
    post:
      operationId: generateChain
      tags: [chains]
      summary: AI-generate a chain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scenarioId:
                  type: integer
                scenarioContext:
                  type: string
                agentIds:
                  type: array
                  items:
                    type: integer
                chainType:
                  type: string
                numSteps:
                  type: integer
                focusArea:
                  type: string
                domainId:
                  type: integer
              required: [chainType, numSteps]
      responses:
        "200":
          description: Generated chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chain"
  /chains/{id}:
    get:
      operationId: getChain
      tags: [chains]
      summary: Get a chain by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chain with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chain"
    put:
      operationId: updateChain
      tags: [chains]
      summary: Update a chain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChainBody"
      responses:
        "200":
          description: Updated chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chain"
    delete:
      operationId: deleteChain
      tags: [chains]
      summary: Delete a chain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Agent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        discipline:
          type: string
        domainId:
          type: integer
        personaDescription:
          type: string
        cognitiveBias:
          type: string
        redTeamFocus:
          type: string
        severityDefault:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        vectorHuman:
          type: integer
        vectorTechnical:
          type: integer
        vectorPhysical:
          type: integer
        vectorFutures:
          type: integer
        isAiGenerated:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, discipline, severityDefault]
    CreateAgentBody:
      type: object
      properties:
        name:
          type: string
        discipline:
          type: string
        domainId:
          type: integer
        personaDescription:
          type: string
        cognitiveBias:
          type: string
        redTeamFocus:
          type: string
        severityDefault:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        vectorHuman:
          type: integer
        vectorTechnical:
          type: integer
        vectorPhysical:
          type: integer
        vectorFutures:
          type: integer
        isAiGenerated:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required: [name, discipline, severityDefault]
    ChainStep:
      type: object
      properties:
        id:
          type: integer
        chainId:
          type: integer
        agentId:
          type: integer
        stepNumber:
          type: integer
        agentLabel:
          type: string
        stepText:
          type: string
        createdAt:
          type: string
      required: [id, chainId, stepNumber, stepText]
    Chain:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        domainId:
          type: integer
        scenarioId:
          type: integer
        threatId:
          type: integer
        isAiGenerated:
          type: boolean
        tags:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ChainStep"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name]
    CreateChainBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        domainId:
          type: integer
        scenarioId:
          type: integer
        threatId:
          type: integer
        isAiGenerated:
          type: boolean
        tags:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: object
            properties:
              agentId:
                type: integer
              stepNumber:
                type: integer
              agentLabel:
                type: string
              stepText:
                type: string
            required: [stepNumber, stepText]
      required: [name]
...