# lib/api-zod/src/generated/types/chainStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WARROOM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChainStep {
  id: number;
  chainId: number;
  agentId?: number;
  stepNumber: number;
  agentLabel?: string;
  stepText: string;
  createdAt?: string;
}