# lib/api-zod/src/generated/types/createAgentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WARROOM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAgentBodySeverityDefault } from "./createAgentBodySeverityDefault";

export interface CreateAgentBody {
  name: string;
  discipline: string;
  domainId?: number;
  personaDescription?: string;
  cognitiveBias?: string;
  redTeamFocus?: string;
  severityDefault: CreateAgentBodySeverityDefault;
  vectorHuman?: number;
  vectorTechnical?: number;
  vectorPhysical?: number;
  vectorFutures?: number;
  isAiGenerated?: boolean;
  tags?: string[];
}